refactor: slide 로직 분리
